Refactor file structure.
