+++ index.js
@@ -1,0 +1,6 @@
+module.exports = {
+  query: require('./lib/query'),
+  prepareQuery: require('./lib/prepare-query'),
+  querySql: require('./lib/query-sql'),
+  prepareSql: require('./lib/prepare-sql')
+}
