+++ lib/detect-fields-type.js
