+++ lib/generateBreaks/normalizeClassBreaks.js
