+++ lib/generateBreaks/utils.js
