+++ lib/geometry/convert-from-esri.js
