+++ lib/geometry/convert-to-esri.js
