+++ lib/geometry/index.js
