+++ lib/geometry/project-coordinates.js
