+++ lib/geometry/reduce-precision.js
