+++ lib/geometry/transform-array.js
