+++ lib/geometry/transform-coordinates.js
