+++ lib/geometry/transform-envelope.js
