+++ lib/geometry/utils.js
