+++ lib/options/index.js
