+++ lib/options/normalizeClassification.js
