+++ lib/options/normalizeOptions.js
