+++ lib/options/normalizeSQL.js
