+++ lib/query-sql.js
@@ -1,0 +1,5 @@
+const alasql = require('./alasql')
+
+module.exports = function (statement, params) {
+  return alasql(statement, params)
+}
