+++ lib/select/compose.js
