+++ lib/select/fields.js
