+++ lib/select/geometry.js
