+++ lib/select/index.js
