+++ package.json
@@ -2,7 +2,7 @@
   "name": "winnow",
   "version": "1.16.13",
   "description": "Apply sql-like filters to GeoJSON",
-  "main": "dist/index.js",
+  "main": "index.js",
   "directories": {
     "test": "test"
   },
